Switch mixed precision to bf16 and harden training shutdown
- Switch fetch_base_trainer, fetch_finetune_trainer, and predict_trainer from
  16-mixed (fp16) to bf16-mixed; bf16's wider dynamic range better suits the
  ALiBi bias, ReZero scaling, and exp candle reconstruction, and drops the fp16
  gradient scaler. run_training now sets torch.set_float32_matmul_precision(
  "high") to enable TF32 for the residual fp32 matmuls.
- Wrap trainer.fit in try/except that drops dataloader refs and runs gc.collect()
  before re-raising, so multiprocessing workers shut down cleanly on Optuna
  pruning / KeyboardInterrupt instead of raising AssertionError in __del__.
- Add CPU-safe tests/test_register.py pinning the precision string for all three
  trainer factories; assert build_split_handlers enables streamer shuffle.

+++ src/ophir/register.py
@@ -126,7 +126,7 @@ def fetch_base_trainer(
 
     trainer = L.Trainer(
         max_steps=max_steps,
-        precision="16-mixed",
+        precision="bf16-mixed",
         default_root_dir=MODEL_DIR,
         accelerator="cuda",
         callbacks=callbacks,
@@ -160,7 +160,7 @@ def fetch_finetune_trainer() -> L.Trainer:
     from lightning.pytorch.loggers import CSVLogger, TensorBoardLogger
 
     trainer = L.Trainer(
-        precision="16-mixed",
+        precision="bf16-mixed",
         max_epochs=10000,
         default_root_dir=MODEL_DIR,
         accelerator="cuda",
@@ -498,7 +498,7 @@ def predict_trainer() -> L.Trainer:
     """
     import lightning as L
 
-    trainer = L.Trainer(precision="16-mixed", accelerator="cuda")
+    trainer = L.Trainer(precision="bf16-mixed", accelerator="cuda")
     return trainer
